crimson/osd, crimson/test: split watch-timeout op into its own TU and add cancel_notify test
WatchTimeoutRequest and Watch::do_watch_timeout() are the only parts of the
watch machinery coupled to the PG operation framework (InternalClientRequest
-> PG::run_executer -> OpsExecuter -> ...). Move them into
watch_timeout_request.cc so watch.cc holds only the connection/notify
bookkeeping and can be compiled and unit-tested without pulling in the whole
OSD. Add a unit-test-only Watch constructor, guarded by a private
unit_test_ctag_t tag and reachable only via create_for_test(), that builds a
Watch with no ObjectContext/PG/connection.

Add a regression test for Watch::cancel_notify()/notify_ack() against a notify
id that is not in in_progress_notifies -- the state hit by the notify-timeout
race, where do_notify_timeout() cancels a watcher whose start_notify() has not
run yet. Pre-fix this asserted (or, under NDEBUG, hit UB); the test target is
built with -UNDEBUG so the old code aborts deterministically instead of hanging.
